test: split observation results validation by concern
